+++ ubo_app/system/bootstrap.py
@@ -90,3 +90,12 @@ def bootstrap() -> None:
         [Path(__file__).parent.joinpath('install_wm8960.sh').as_posix()],  # noqa: S603
         check=True,
     )
+
+
+def install_docker() -> None:
+    """Install Docker."""
+    # Run the install_docker.sh script
+    subprocess.run(
+        [Path(__file__).parent.joinpath('install_docker.sh').as_posix()],  # noqa: S603
+        check=False,
+    )