+++ ubo_app/utils/loop.py
@@ -3,15 +3,20 @@ from __future__ import annotations
 
 import asyncio
 import threading
-from typing import TYPE_CHECKING, Coroutine
+from typing import TYPE_CHECKING, Callable, Coroutine, TypeVarTuple, Unpack
 
 from redux import FinishEvent
+from typing_extensions import TypeVar
 
 from ubo_app.constants import DEBUG_MODE
 from ubo_app.store import subscribe_event
 
 if TYPE_CHECKING:
-    from asyncio import Handle
+    from asyncio import Future, Handle
+
+
+T = TypeVar('T', infer_variance=True)
+Ts = TypeVarTuple('Ts')
 
 
 class WorkerThread(threading.Thread):
@@ -30,6 +35,14 @@ class WorkerThread(threading.Thread):
     def run_task(self: WorkerThread, task: Coroutine) -> Handle:
         return self.loop.call_soon_threadsafe(self.loop.create_task, task)
 
+    def run_in_executor(
+        self: WorkerThread,
+        executor: object,
+        task: Callable[[Unpack[Ts]], T],
+        *args: Unpack[Ts],
+    ) -> Future[T]:
+        return self.loop.run_in_executor(executor, task, *args)
+
     def stop(self: WorkerThread) -> None:
         self.loop.call_soon_threadsafe(self.loop.stop)
 
@@ -40,3 +53,14 @@ thread.start()
 
 def _create_task(task: Coroutine) -> Handle:
     return thread.run_task(task)
+
+
+def _run_in_executor(
+    executer: object,
+    task: Callable[[Unpack[Ts]], T],
+    *args: Unpack[Ts],
+) -> Future[T]:
+    return thread.run_in_executor(executer, task, *args)
+
+
+_ = _create_task, _run_in_executor